Added Identicons as default avatar

+++ applications/commands/profile/avatar.php
@@ -25,9 +25,9 @@ if (isset($_POST['edit'])) {
             'image/gif'
         );
 
-        $gravatar = (isset($_POST['gravatar'])) ? '1' : '0';
+        $avatar = (isset($_POST['avatar'])) ? $_POST['avatar'] : '';
 
-        if ($gravatar == "1") {
+        if ($avatar == 1) { // Using Gravatar
             $MYSQL->bind('id', $IKO->sess->data['id']);
             if ($MYSQL->query("UPDATE {prefix}users SET avatar_type = 1 WHERE id = :id") > 0) {
                 $notice .= $IKO->tpl->entity(
@@ -39,41 +39,70 @@ if (isset($_POST['edit'])) {
                 throw new Exception ($LANG['bb']['profile']['error_adding_gravatar']);
             }
 
-        } elseif (!$_FILES['avatar']) {
-            throw new Exception ($LANG['global_form_process']['all_fields_required']);
-        } elseif (!in_array($_FILES['avatar']['type'], $mime)) {
-            throw new Exception ($LANG['global_form_process']['invalid_file_format']);
-        } else {
+        } elseif ($avatar == 2) { // Using identicons
+            require_once('applications/libraries/Identicon/Generator/BaseGenerator.php');
+            require_once('applications/libraries/Identicon/Generator/GeneratorInterface.php');
+            require_once('applications/libraries/Identicon/Generator/GdGenerator.php');
+            require_once('applications/libraries/Identicon/Identicon.php');
 
-            $image = $_FILES['avatar'];
-            $bin_dir = 'public/img/bin/' . $IKO->sess->data['id'] . '.png';
-            copy($image['tmp_name'], $bin_dir);
-            list($width, $height, $type, $attr) = getimagesize($bin_dir);
+            $avatar_dir = 'public/img/avatars/' . $IKO->sess->data['id'] . '.png';
+            $identicon = new \Identicon\Identicon();
+            $imageData = $identicon->getImageData($IKO->sess->data['username'], 500);
+            $image = imagecreatefromstring($imageData);
+            if ($image !== false) {
+                imagepng($image, $avatar_dir);
+                imagedestroy($image);
+                $MYSQL->bind('user_avatar', $IKO->sess->data['id'] . '.png');
+                $MYSQL->bind('id', $IKO->sess->data['id']);
+                $MYSQL->query("UPDATE {prefix}users SET user_avatar = :user_avatar, avatar_type = 2 WHERE id = :id");
+                $notice .= $IKO->tpl->entity(
+                    'success_notice',
+                    'content',
+                    $LANG['bb']['profile']['successful_identicon']
+                );
+            } else {
+                throw new Exception($LANG['errors']['generate_identicon']);
+            }
 
-            if ($width > 500 && $height > 500) {
-                throw new Exception ($LANG['global_form_process']['img_dimension_limit']);
+        } elseif ($avatar == 0) { // Uploading regular avatar
+            if (!$_FILES['avatar']) {
+                throw new Exception ($LANG['global_form_process']['all_fields_required']);
+            } elseif (!in_array($_FILES['avatar']['type'], $mime)) {
+                throw new Exception ($LANG['global_form_process']['invalid_file_format']);
             } else {
 
-                unlink($bin_dir);
-                $avatar_dir = 'public/img/avatars/' . $IKO->sess->data['id'] . '.png';
-                if (copy($image['tmp_name'], $avatar_dir)) {
-                    $MYSQL->bind('user_avatar', $IKO->sess->data['id'] . '.png');
-                    $MYSQL->bind('id', $IKO->sess->data['id']);
-                    $MYSQL->query("UPDATE {prefix}users SET user_avatar = :user_avatar WHERE id = :id");
-                    $MYSQL->bind('id', $IKO->sess->data['id']);
-                    $MYSQL->query("UPDATE {prefix}users SET avatar_type = 0 WHERE id = :id");
-                    $notice .= $IKO->tpl->entity(
-                        'success_notice',
-                        'content',
-                        $LANG['bb']['profile']['successful_upload_avatar']
-                    );
+                $image = $_FILES['avatar'];
+                $bin_dir = 'public/img/bin/' . $IKO->sess->data['id'] . '.png';
+                copy($image['tmp_name'], $bin_dir);
+                list($width, $height, $type, $attr) = getimagesize($bin_dir);
 
+                if ($width > 500 && $height > 500) {
+                    throw new Exception ($LANG['global_form_process']['img_dimension_limit']);
                 } else {
-                    throw new Exception ($LANG['bb']['profile']['error_upload_avatar']);
+
+                    unlink($bin_dir);
+                    $avatar_dir = 'public/img/avatars/' . $IKO->sess->data['id'] . '.png';
+                    if (copy($image['tmp_name'], $avatar_dir)) {
+                        $MYSQL->bind('user_avatar', $IKO->sess->data['id'] . '.png');
+                        $MYSQL->bind('id', $IKO->sess->data['id']);
+                        $MYSQL->query("UPDATE {prefix}users SET user_avatar = :user_avatar WHERE id = :id");
+                        $MYSQL->bind('id', $IKO->sess->data['id']);
+                        $MYSQL->query("UPDATE {prefix}users SET avatar_type = 0 WHERE id = :id");
+                        $notice .= $IKO->tpl->entity(
+                            'success_notice',
+                            'content',
+                            $LANG['bb']['profile']['successful_upload_avatar']
+                        );
+
+                    } else {
+                        throw new Exception ($LANG['bb']['profile']['error_upload_avatar']);
+                    }
+
                 }
 
             }
-
+        } else {
+            throw new Exception ('Test');
         }
 
     } catch (Exception $e) {
@@ -100,20 +129,28 @@ $IKO->tpl->addBreadcrumb(
     '#',
     true
 );
-$bc = $IKO->tpl->breadcrumbs();
+$breadcrumbs = $IKO->tpl->breadcrumbs();
 
+$avatar_checked = ($IKO->sess->data['avatar_type'] == "0") ? ' checked' : '';
 $gravatar_checked = ($IKO->sess->data['avatar_type'] == "1") ? ' checked' : '';
+$identicon_checked = ($IKO->sess->data['avatar_type'] == "2") ? ' checked' : '';
+
 $content .= '<form id="tango_form" action="" method="POST" enctype="multipart/form-data">
+                 <fieldset>
+                 <input type="radio" id="avatar" name="avatar" class="avatar_activator" value="0"' . $avatar_checked . ' /> <label for="avatar">' . $LANG['bb']['profile']['use_avatar'] . '</label>
                  <div class="iko avatar_uploader">
                    <label for="avatar">' . $LANG['bb']['profile']['change_avatar'] . '</label>
                    <input type="file" name="avatar" id="avatar" />
-                   <br />
                  </div>
-                 <input type="checkbox" id="gravatar" name="gravatar" value="1"' . $gravatar_checked . ' /> <label for="gravatar">' . $LANG['bb']['profile']['use_gravatar'] . '</label>
+                 <br />
+                 <input type="radio" id="gravatar" name="avatar" value="1"' . $gravatar_checked . ' /> <label for="gravatar">' . $LANG['bb']['profile']['use_gravatar'] . '</label>
+                 <br />
+                 <input type="radio" id="identicon" name="avatar" value="2"' . $identicon_checked . ' /> <label for="identicon">' . $LANG['bb']['profile']['use_identicon'] . '</label>
+                 </fieldset>
                  <br /><br />
                  <input type="submit" name="edit" value="' . $LANG['bb']['profile']['form_save'] . '" />
                </form>';
 
-$content = $bc . $notice . $content;
+$content = $breadcrumbs . $notice . $content;
 
 ?>
